pyproject.toml
[project]
name = "im-bot-email"
version = "0.1.0"
description = "Email bot — listens for new emails via IMAP IDLE, parses them into tasks, executes, and replies via SMTP."
requires-python = ">=3.10"
dependencies = [
    "imapclient>=3.0.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/im_bot_email"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
